+++ src/FarNet.Redis/DB.cs
@@ -1,4 +1,4 @@
-﻿
+﻿using System;
 using System.Collections.Concurrent;
 using System.Linq;
 
@@ -6,7 +6,21 @@ namespace StackExchange.Redis;
 
 public static class DB
 {
-    static readonly ConcurrentDictionary<string, IDatabase> _db = [];
+    private static readonly ConcurrentDictionary<string, IDatabase> s_databases = [];
+    private static IDatabase? s_database;
+
+    public static IDatabase? DefaultDatabase => s_database;
+
+    public static IDatabase OpenDefaultDatabase()
+    {
+        if (s_database is { })
+            return s_database;
+
+        var configuration = Environment.GetEnvironmentVariable("FARNET_REDIS_CONFIGURATION")
+            ?? throw new InvalidOperationException("Requires environment variable FARNET_REDIS_CONFIGURATION.");
+
+        return s_database = Open(configuration);
+    }
 
     public static IDatabase Open(string configuration)
     {
@@ -17,7 +31,7 @@ public static class DB
     public static IDatabase Open(ConfigurationOptions options)
     {
         var key = options.ToString();
-        var db = _db.GetOrAdd(key, key =>
+        var db = s_databases.GetOrAdd(key, key =>
         {
             var redis = ConnectionMultiplexer.Connect(options);
             return redis.GetDatabase();
@@ -31,17 +45,20 @@ public static class DB
         var options = ConfigurationOptions.Parse(configuration);
 
         var key = options.ToString();
-        if (_db.TryRemove(key, out IDatabase? db))
+        if (s_databases.TryRemove(key, out IDatabase? db))
             db.Multiplexer.Close();
     }
 
     public static void Close(IDatabase db)
     {
+        if (ReferenceEquals(s_database, db))
+            s_database = null;
+
         db.Multiplexer.Close();
-        foreach (var kv in _db.ToList())
+        foreach (var kv in s_databases.ToList())
         {
             if (ReferenceEquals(kv.Value, db))
-                _db.TryRemove(kv.Key, out _);
+                s_databases.TryRemove(kv.Key, out _);
         }
     }
 }